AppControl Manager v.2.0.1.0
You can now launch as many instances of the AppControl Manager app as you want, they are able to work side by side. This improves the scalability of the AppControl Manager and offers isolation scenarios. For example, you can launch one instance of the app with standard user privileges and 2 instance with elevated permissions and run tasks simultaneously.

Updated multiple NuGet dependencies to the latest versions.

Added a new section to system information to view and delete Intune App Control policies. Now you can instantly view the App Control policies that you deploy to your Intune, right inside of the app and remove existing ones. The app also implements advanced details of each policy that it deploys to the Intune, allowing you to view details of the Intune deployed policies just like the locally deployed policies. For this reason, it is always recommended to use the AppControl Manager to deploy your App Control policies to the Intune.

Added "Open in Policy Editor" option to the Create Supplemental policy and Deny policy pages. At the end of the operations in those pages, you will have the option to open the created policy in Policy Editor and inspect it in details or fine tune it.

Added advanced centralized account management to the pages that offer signing into Microsoft tenants. It offers the following features:

1- Log into as many tenants as you want.

2- View the list of all of the accounts you've signed into.

3- Completely and instantly remove any of the previously authentication accounts from the app and its memory.

4- The ability to switch between authentication accounts for the operations in each page.

5- The ability to choose the sign in experience between using the system browser or using a more native method via Web Account Manager (native Windows feature).

Significantly improved the ability of the AppControl Manager to be translated to other languages by extracting and separating the string texts from the code sections.

Fixed multiple typos in the UI elements' texts.

Changed the style of the page titles, they are now more consistent, translatable to other languages and their colors are no longer affected by the current OS accent color. This can improve readability in case your OS accent color is set to a color that is hard to read.

Added the ability to change the application's language from the app's settings page. It supports only English and Hebrew at the moment but more will be added in the future. Please refer to the GitHub discussions post to find out how to contribute in translating the AppControl Manager into your own native language. You will need to restart the app for the language changes to fully take effect.

Added the ability to change the UI (User Interface) direction. You can now switch between Left-To-Right and Right-To-Left. The changes will be instantly visible in the entire app.