AppControl Manager v.2.0.11.0 (#683)
Fixed a visual bug where the text in the Deployment page would show "local deployment" instead of "Online deployment". Thanks @zygmuntszpak for reporting it.

When deploying policies via AppControl Manager, you can now select more than 1 Intune or M365 groups to be assigned to the Intune policy. Previously, you could only select 1 group. For this reason, the combobox design has evolved to a multi-select ListView element in the Deployment page. Thanks @zygmuntszpak for the suggestion.

Updated NuGet dependencies to the latest versions.

The update also includes packageQuery permission in the manifest for the next versions to switch to that API.