Switching from PS to Rust for a WMI call in AppControl Manager (#690)
The goal is to completely avoid calling Windows PowerShell in the AppControl Manager because it makes the app depend on it. This is the first PR that addresses one of the few cases where Device Guard WMI data used to get retrieved via PowerShell but now a completely self-contained Rust based solution is used. One of the benefits of this approach is that the app will work normally on systems where Windows PowerShell (the built-in one) is not allowed to be used. This is also much faster than using PowerShell cmdlets.

bumped app version to 2.0.12.0