Harden Windows Security v0.7.5.1 (#691)
The Harden Windows Security module is now resilient. Whenever an error occurs, it will display the full error message in a modern window but will not terminate the session, instead will gracefully handle the situation, allowing you to continue past the error and perform other tasks.

Updated the Microsoft DLL dependencies to the latest versions.

Added a new button to the AppControl Manager tab that will open the Microsoft Store to install it. This is a new option in addition to the existing option that installs AppControl Manager from the GitHub source.

improved the installation process for AppControl Manager when using GitHub source. It uses MSIXBundle now instead of the MSIX file, making it compatible with ARM64 machines.

Added OS version to the error message that is displayed whenever there is an error in the Harden Windows Security module.

Added 2 new policies to the Windows Networking category, NTLM Sub-category. It now has the following policies: - Thanks @Mrkuff for the suggestion and research

For SMB.

For all incoming connections.

For all outgoing connections. - This can prevent you from using RDP (Remote Desktop) remotely via IP address which is insecure as it needs public exposed ports and uses NTLM. You can use Quick Assist or Bastion for Azure VMs which are more secure alternatives. Local RDP such as for Hyper-V enhanced session is not affected.

Disables the RPC Endpoint Mapper Client Authentication policy. It is recommended to be disabled when NTLM is completely blocked.

This policy used to be in the Miscellaneous Category and was set to enabled, but it is now part of this section and is set to disabled. The reason is explained in the Microsoft Document. There is no need for any changes to your current policies but if you use the NTLM blocking sub-category, especially after this update, it needs to be disabled and this additional policy will do exactly that to be according to the Microsoft's recommendations.
Removed the policy from the Windows Firewall category that would disable connections for the Domain profile. This change is applied so that the module will be more compatible on Windows Servers.

Added a new policy to the Windows Firewall category that will show notifications for the Domain profile.