Partial C++ rewrite and improved UI translation (#695)
Rewrite of parts of the AppControl Manager in Rust and C++. Parts where it made sense to use a lower level language were involved in this rewrite. This change also made the application more self-sufficient than before by relying on itself for features such as native COM usage. Now it's faster and compatible with Native AOT nature of the app.

Improved UI translation.