AppControl Manager v2.0.13.0 (#701)
Changed the progress bar to a progress ring in the Simulation page. It takes less space in the UI and offers the same functionality.

Also in the Simulation page, the "Save output to CSV" button's behavior has changed. It used to be a toggle button which would only export the results at the end of the simulation but now it's a regular button and you can use it at any time to export the results to a CSV file.

The Simulation page also features a new info bar that will appear giving you real time info about its operations. If you forgot to select an XML file or browse for files to participate in the simulation, that info bar will be used to communicate the right message to the user in a more subtle way rather than displaying error.

The length of all of the IDs in the generated policies have been shortened.

Fixed an arithmetic overflow issue.

Reimplemented the sorting features in all List Views. Removed the dedicated Sort button from the pages and instead the column headers of the List Views are clickable now, allowing you to instantly and easily sort the data based on that column's data. This makes sorting decisions faster as you don't have to look through a list. Now you just click/tap on the header of each column and it will be sorted. The first tap/click will sort that column in Descending order, the 2nd tap/click will sort it in Ascending order.