AppControl Manager v2.0.14.0
AppControl Manager can now open App Control Policy XML files directly from the File Explorer. It will appear as an option in the "Open With" right-click context menu. You can optionally configure it as the default handler for XML files which will let you easily double-click/tap on an App Control XML file and open it in AppControl Manager's Policy Editor.

If the location where the file is located requires elevated permissions for write access then AppControl Manager will request Administrator privileges (you will see a UAC prompt), otherwise it will launch with standard privileges.
Changed the source of the Microsoft Recommended Driver block rules from the Microsoft Learn page's markdown to the zip file that Microsoft provides.

Added an Info bar to the Policy Editor page that shows the status of the running operations.

The Supplemental policy creation page has a new feature that allows you switch between different modes. The default mode is the same as before which is creating new supplemental policies, and there is a new mode that allows you to select an existing App Control policy and add any rules you create in the page to it. You can use this new mode to easily and quickly expand any of your existing App Control policies.