AppControl Manager v2.0.14.0 continuation (#708)
Created a new C++ based program dedicated to managing every aspect of scheduled tasks. It will replace any calls to PowerShell cmdlets in the AppControl Manager and will later be used in the Harden Windows Security application as well. It is self-contained, small (300KB) and fast.

When sorting or searching through the data in the ListViews, the horizontal scrolling position will be maintained and will no longer be reset to the left side, improving the user experience.

All of the content dialogs in the app are now theme-aware and will react to the dark/light theme changes.

Improved text translations for other languages.

Code refactor and unification.

Improved Microsoft Learn links in the comments so they are no longer language specific and will change according to your language.

Added extra accessibility texts for various elements.

Internalized Windows Service management. The application no longer uses the sc.exe, reducing dependency on executables outside of the AppControl Manager package.