Implemented CIP to XML and back, improved Policy Editor
AppControl Manager now enables all users who are on the Home edition of the Windows OS to be able to use all of the advanced features offered by App Control. Since Home edition doesn't have the necessary PowerShell cmdlets and AppControl Manager does not rely on PowerShell at all, it can work flawlessly on that edition.

You can now accurately and systematically convert any old and new CIP binary files back to XML files. CIP binaries are App Control policies that are consumed by the Code Integrity in the Windows OS. I wrote a thorough document about the reverse mechanism in here if you are interested.

The Policy Editor Page now accepts both XML and CIP binary files. If you load a CIP file and then use the Save button, a new XML file will be created with the same name in the AppControl Manager directory (if running elevated) or in the same location as the CIP file is (if unelevated).

The AppControl Manager now shows a small modern badge on the taskbar icon whenever it's performing a long running activity such as performing simulation or during local file scans, giving you more information.

After installed AppControl Manager, you will be able to simply double-click/tap on CIP binary files and instantly decode and open them in the Policy Editor page. You can also do the same with XML files via right-click and selecting AppControl Manager in the "Open With..." context menu. The CIP and XML files will have a distinct beautiful icon when AppControl Manager is set as the default handler for them.