AppControl Manager v2.0.20.0 (#732)
Made the Update page available for when you install the AppControl Manager from the Microsoft Store. Now it's always available no matter the source of the installation, however its content and available options are dynamically changed. If you installed it from Microsoft Store and a new version is available, the update button will open the mini version of the Microsoft Store where you can press the update button to install the new version.

When installed from the Microsoft Store, a new section becomes available in the Update page where you will be taken to the mini store page to Rate and optionally review the AppControl Manager.

When installed from the Microsoft Store, the Update page becomes available to both elevated and unelevated sessions.

Added full support for a 3rd language: Greek

Adjusted the sidebar elements' orientations to utilize the space better.

A lot of new texts have been translated to other non-English languages, mostly error and operational messages that you see in the logs or info bars throughout the app.

Created a new thorough high-performance Rust-based analyzer for the resource files, ensuring the existing and future multi-lingual support remains error-free and compliant by catching any possible mistakes early. It will also be added as a separate job to the GitHub action.

Added a new section to the Settings page where you can provide a custom override path for the Code Integrity schema XSD file used by the application. This is mostly intended for users of Home edition of Windows where that file doesn't exist by default.