Fixed an issue with Add to policy feature in certain pages (#747)
When trying to add rules to an existing policy in the Event logs or MDE Advanced hunting pages, it would fail because all rule options from the new policy were being removed. Removed that step because it's unnecessary as the merge logic already ensures the user selected policy's rule options will be honored in the final policy. This change also fixes the related error reported in the issue AppControl Manager latest update 2.0.21.0 Bug? #746

Bumped app version to 2.0.23.0

Updated a NuGet dependency to the latest version.