AppControl Manager v2.0.24.0 (#753)
Updated dependencies to the latest version.

Animated the shadow in the Supplemental policy creation page to react to hover/touch.

Added country flags to the languages selection drop down menu in the Settings page.

Significantly improved the appearance and design of the scalability radial gauges in various pages that implement it.

Natively implemented package signing logic to avoid the need for SignTool.exe. This feature is only used when the app is installed from the GitHub source and you update to the latest version.

Improved localization.