AppControl Manager v2.0.26.0 (#757)
On Copilot+ PCs, you will be able to use Recall to navigate to the different pages you were using at specific points in time in AppControl Manager.

Added a new option in the app's Settings page called Screen Shield. It prevents other apps such as screen recorders from recording or taking screenshots of the window of the AppControl Manager. This option also prevents Recall from taking screenshots of the app.

You can now change the app's language without the need to restart the app. Every text on the UI will instantly change to the language you select at runtime.

Added smooth animation to all of the Info Bars throughout the app when they transition between open and closed states.

Added beautiful shadow to all of the Content Dialogs throughout the app to enhance the visual depth and focus on the content.

Fixed the texts on the radial gauges not being visible in light theme. They are now aware of the current app theme.

Fixed the Info Bar in the "Build New Certificate" page, it didn't have correct bindings for its open/closed states.

Fixed an issue in the "System Information" page where the text showing the type of the policy you're trying to swap to wouldn't be visible on the Content Dialog.

Added new RMM products to the RMM blocking policy.

Improved memory management by turning off page caching for multiple additional pages.

Relocated all the LibraryImports to the NativeMethods class to follow best practices, improve performance and organize the code.

Updated the NuGet dependencies to the latest versions.

Improved build process.