AppControl Manager v2.0.31.1
Added WHQL File Publisher scan level support to the AppControl Manager. This is the most secure scan level for signed files and is now the default level for all file scans. It has automatic fallback to File Publisher, Publisher and Hash levels so it can create rule for all types of files.

Added File Path and Wildcard Folder Path levels to the Allow New Apps page.

Added security score to all of the scan levels in the app so you can quickly see the security each level offers.

Updated the icons for AppControl Manager, they now have less unused spaces around them.

Created a Rust-based project that includes all of the native Interops and COM usages, replacing the CsWin32 implementations with more high performance code. Any future native code will be included in it.

Improved the RTL (Right-To-Left) layout and design by making the entire title bar, including the Minimize/Maximize/Close buttons, RTL aware. This means the title bar will be on the right side of the window when the app is in RTL mode. You can configure RTL/LTR mode in the app settings via a toggle switch.

Added the 6th Sidebar button for Supplemental Policy page.

AppControl Manager version bump to 2.0.31.0.

Fixed an issue with the Policy Editor now saving the policy name and policy ID when you press the Save button.

Updated NuGet dependencies to the latest version.