More improvements for v2.0.31.0
In the Get Code Integrity File Hashes page, added 2 new features:

You can now right-click or Tap + Hold on the Browse button to view the selected file.

Added animated icon to the Clear Data button.

Sub-pages now appear in the search results when you try to find a page using the search bar at the top of the app's user interface, previously only the main pages would appear in the search results.

Improved localization.

When you right-click or Tap + Hold on rows of the List Views that show scan results, you now have an option called "Open in File Explorer". When you select that, the directory that contains the file will be opened in File Explorer, allowing you to interact with it quickly without the need to manually browse the directories to find that file.

It also has a keyboard shortcut associated with it: CTRL + O (Control button + O as in Orange)
Added additional metadata to the main Window of the app for screen readers to use for accessibility purposes.

After creating a policy in the app, you now have a new option that will let you open the policy file using the default file handler set in the OS, such as Visual Studio Code or Notepad. This is in addition to the existing option where you could open the policy in the Policy Editor with 1 click/tap on a button. As a result of this change, the button has changed into a drop-down button to be able to show more options.

The scan level of the Kernel-mode Supplemental policy creation has been changed from FilePublisher to WHQLFilePublisher.