AppControl Manager v2.0.31.0 (#774)
Added WHQL File Publisher scan level support to the AppControl Manager. This is the most secure scan level for signed files and is now the default level for all file scans. It has automatic fallback to File Publisher, Publisher and Hash levels so it can create rule for all types of files.

Added File Path and Wildcard Folder Path levels to the Allow New Apps page.

Added security score to all of the scan levels in the app so you can quickly see the security each level offers.

Updated the icons for AppControl Manager, they now have less unused spaces around them.

Created a Rust-based project that includes all of the native Interops and COM usages, replacing the CsWin32 implementations with more high performance code. Any future native code will be included in it.

Improved the RTL (Right-To-Left) layout and design by making the entire title bar, including the Minimize/Maximize/Close buttons, RTL aware. This means the title bar will be on the right side of the window when the app is in RTL mode. You can configure RTL/LTR mode in the app settings via a toggle switch.

Added the 6th Sidebar button for Supplemental Policy page.

AppControl Manager version bump to 2.0.31.0.

Fixed an issue with the Policy Editor not saving the policy name and policy ID when you press the Save button. - Thanks @Minifab and @Obscurity1935 for reporting it. 🙏🏻

Updated NuGet dependencies to the latest version.

In the Get Code Integrity File Hashes page, added 2 new features:

You can now right-click or Tap + Hold on the Browse button to view the selected file.

Added animated icon to the Clear Data button.

Sub-pages now appear in the search results when you try to find a page using the search bar at the top of the app's user interface, previously only the main pages would appear in the search results.

Improved localization.

When you right-click or Tap + Hold on rows of the List Views that show scan results, you now have an option called "Open in File Explorer". When you select that, the directory that contains the file will be opened in File Explorer, allowing you to interact with it quickly without the need to manually browse the directories to find that file.

It also has a keyboard shortcut associated with it: CTRL + O (Control button + O as in Orange)
Added additional metadata to the main Window of the app for screen readers to use for accessibility purposes.

After creating a policy in the app, you now have a new option that will let you open the policy file using the default file handler set in the OS, such as Visual Studio Code or Notepad. This is in addition to the existing option where you could open the policy in the Policy Editor with 1 click/tap on a button. As a result of this change, the button has changed into a drop-down button to be able to show more options.

The scan level of the Kernel-mode Supplemental policy creation has been changed from FilePublisher to WHQLFilePublisher.

Added "Configure Rule Options" quick action for when a policy is created in the "Create Policy" page.

Updated the icon of the "Get Secure Policy Settings" page in the main navigation.