AppControl Manager v2.0.33.0
Updated the Rust build processes for native library generation and updated the package generation scripts accordingly.

This change improves the security and performance of calls to the Rust code as they will be absorbed in the app's executable rather than being a separate file.

You can read more about it here: https://learn.microsoft.com/en-us/dotnet/core/deploying/native-aot/interop

The build script also has a new parameter for installing the AppControl Manager right after building it from the source code.

When you right-click or hold + tap on an entry in a List View and select to open it in file explorer, the file will be scrolled to if it's down in the list and it will be highlighted, allowing you to quickly locate it.

Updated NuGet dependencies.

When you enable Sound in the app settings (it is off by default), there will be sound effects for typing in the text fields in the app.