Harden Windows Security App core functionality (#798)
Harden Windows Security Core Functionality
This PR implements the core functionality of the new upcoming Harden Windows Security app. It focuses on low level implementation of the security features that will be available in the app. The goal is to provide a solid foundation for the app to build upon, ensuring that it can effectively and independently manage and enhance the security posture of Windows systems.

So far the following categories have been completely implemented:

Microsoft Defender
Attack Surface Reduction
Optional Features and Capabilities
There is also a new page for viewing all of the Group Policies applied on the system, exporting them to JSON, converting POL files to JSON and vise versa. It's just the beginning and so many other features will be added to this page.

Many features have been implemented in both Rust and C# and whichever version that will be selected will be based on different criteria such as performance and other implementation complexities surrounding packaged self contained apps.

Everything implemented so far in the Harden Windows Security app is fully Native AOT compatible and it will stay that way, ensuring a small package and very fast startup times.

Implementing the rest of the categories is rather straightforward from here on out, so expect more PRs in the future.

AppControl Manager's code base have been modified to provide interoperability with the Harden Windows Security app. This reduces code duplication and allows for a more unified approach to managing security features across the system. Many methods have been refined and generalized to work with both apps seamlessly. They can be converted into a shared library project in the future but that's not a priority at the moment. Conditional compilation is used to ensure that the code works correctly in both contexts.

All of the dependencies for AppControl Manager and the Harden Windows Security app have been updated to their latest versions.

I could continue working on the PR till the Harden Windows Security development is complete, without merging this PR, but I need to release a new AppControl Manager version with updated dependencies to ensure it contains the latest security patches and improvements.

If you're really enthusiastic and want to give super early feedback for the upcoming Harden Windows Security app, you will need to build it from the source code for now.