Harden System Security v1.0.2.0 (#811)
Added an app alias for the Harden System Security. You can now launch it from commandline via HSS.exe.

Added a new file association for .pol files. The app will now handle files with the extension .pol and open them in the Harden System Security application's Group Policy Editor page. They will also have a beautiful icon in the file explorer. By default, .pol files don't have any icons associated with them.

When you hover over the link button for each security measure, a tiny flyout will appear that'll open the URL of the security measure in a mini browser, allowing you to quickly browse the resource and get accurate up to date information on demand. You have the option to disable this feature in the app settings.

Reduced the amount of empty spaces around each security measure item in the List Views by better utilizing the available space.

Reduced the amount of empty spaces in each page between the page header and the Guide button. (This change has been applied to the AppControl Manager as well.)

Improved some of the security measure names in the TLS category to be more descriptive and easier to understand.

Made the columns in the List View in the Group Policy Editor page resizable by offering draggable areas in the header.

Added a new setting to the app settings page that allows you to control whether the List Views that show items in groups have sticky header or the header will scroll with the rest of the items.

Fixed a visual issue with animations in some of the pages' toolbars. During back navigation (when using the back button), some of them would have half opacity.

The settings page now shows the correct app name in the About section.