Complete Documentation and Readme Design Update (#831)
Complete Documentation and Readme Design Update
This PR aligns many of the repository's documents with the recent changes, especially the release of the new Harden System Security app.

The most notable change is moving the Security Measures Categories to the Wiki, each having their own dedicated pages.

One of the community members started the conversation about this change here: Seeking the way forward (Harden Windows Security Module / AppControl Manager) #775

Lots of visual upgrades.

Links to Harden Windows Security module has been removed from the Readme and Moved to its own page in the Wiki. This is because new or existing users are encouraged to use the new Harden System Security app. There are about 15% of the features of the module still not implemented in the new app. They'll be implemented very soon.

When I first created this repo, it was way too simple so I included everything in the Readme back then, but as I continued adding more and more content, it became necessary to rethink the whole design and restructure the Readme page so that the new or returning users would have easier and better time consuming the documents and offerings of this repository.

No information was removed during this transition other than those that no longer apply or were obsolete.

The Rationale page has been merged with the Readme, so the content of the Rationale page is now included in Readme page. The Rationale page continues to exist here for backward compatibility for some time.

Let me know in the discussions or comment section if you have any feedback or suggestions!