Update Harden System Security.md
What's New

You can now export all Audit Policies from a system to a file after retrieving them. Additionally, you can import this file back into the app to automatically apply the same policies on another system. This allows you to create a custom Audit Policies template once and effortlessly re-use it across multiple systems using the Harden System Security app. Having such ability has been one of the community's top requests. This behavior will be expanded to more sections of the app in the near future.

Fixed an issue in Audit Policies retrieval, it now displays the correct number of audit policies that are available.

Updated the More Info link in the Update page when checking for update, it now points to the Harden System Security app's home page in the repository wiki.

Added webpage reference URLs for some the security measures that didn't have it.

Added icons to the buttons in the UI that didn't have it. Now all of the buttons have icons for better visual representation.

The Optional Windows Features page now displays the progress with more details, including the current action taking place and the name of the item. The same data are logged too.

The Optional Windows Features page now has 3 additional buttons for applying/removing/verifying recommended configurations. The page has a Retrieve button that shows over 500 items. The 3 new buttons have specific tasks according to the document linked above where they only enable/disable certain features and capabilities in the OS for optimal security. This is the same feature that the module has.

All texts displayed in the Info Bars in Harden System Security (and AppControl Manager) are now selectable.

Starting with this update, in the Optional Windows Features page, the MDAG (Microsoft Defender Application Guard) is only applied/verified/removed if OS is older than 24H2. That's because it's deprecated and no longer available in the OS since 24H2.

The GitHub documentation page now points to the Harden System Security Home page instead of the GitHub readme. This is consistent with the behavior in the AppControl Manager app.

Improved the Logs page in both Harden System Security and AppControl Manager by making it easier to select multiple lines at the same time. Additionally, there is a new button for copying the selected logs to clipboard. Also, the combobox no longer shows the full file path of the log file, instead it only shows the log file name. There is already a button that opens the logs folder so showing full log file path is not necessary and would take too much unnecessary space in the UI.

You can now export the certificates in the Certificate Checking page to JSON via the new button added in this update.

Cleaned up unused dependencies in both apps' projects.

Added security flags to the C++ projects.