Harden System Security v1.0.8.0 (#834)
Updated dependencies to the latest versions.

Added a new section to the BitLocker page called "Management".

It allows you to view all of the details of each volume and see if any of them is BitLocker encrypted.
You can also view key protectors of the unlocked drives and back them up for safe keeping.
There is also an option to delete any existing key protector.
Sidebar is resizable.
Tip

This section is going to get a lot more features in the next update.