Officially deprecating Harden Windows Security PowerShell Module (#841)
All cmdlets now display a warning with links to Microsoft Store and GitHub document to guide user to install and use the new Harden System Security app.

Added a notice to the GUI at top about deprecation with hyperlinks that point to the Microsoft Store page and GitHub document of the new Harden System Security app.

The Unprotect-WindowsSecurity and Confirm-SystemCompliance cmdlets only show the deprecation warning with links, and do not process anything else. The GUI experience still offers their tasks.

The new Harden System Security app is at a mature and fully featured state and is ready to fully replace the module.

Removed the auto update check from the module.