AppControl Manager v2.0.39.0
What's New

Improved memory and resource management throughout the app.

Enabled extensive analyzers that run during compilation and are able to detect potential issues with broader visibility and accuracy. They increase compilation time (aka building) of the app package, but are worth it.

Created advanced custom incremental collections for list views. They now load items in chunks as you scroll instead of loading everything at once, which improve performance significantly and reduce memory usage when dealing with large datasets such as millions of items in scan results. The distinction between a standard and an incremental collection is subtle from the user's perspective because now the app detects ahead of time when you are likely to scroll down so it will load the next chunk of items before you actually reach the end of the current chunk. This way, it feels like everything is loaded at once, but in reality, it is not. The app will only load what is necessary to display at any given time. At moment, this feature is only implemented for the Create Deny Policy page's Scan Results list view. Over time, it will be implemented in other list views that can potentially have a large number of items. You can try it out and give me feedback.

In the UI, you can view the total count of all of the items vs the count of the items currently loaded.
Added a new setting to the Settings page of the app, under the "Behavior" section, to enable or disable auto column resizing for certain List Views that support it. It is turned off by default so you can turn it on and give it a try if you want. It automatically resizes the columns as you scroll to fit the visible content. The columns will expand and contract dynamically in real time.

Fixed a small issue in the Create Deny Policy page where a warning on the UI for the "Package Family Name" and "Pattern-Based" sections was showing up on the "Files and Folders" section's Info Bar instead.

Improved the right-click experience for many List Views:

If you right-click an item that is already selected (even when multiple are selected), it now keeps the current selection and just shows the context menu for that selection.

If you right-click an unselected item, it now clears the previous selection(s) and selects only that item.

You can now change the font family of the data displayed in various List Views in the app in the Settings page. Currently applies the following pages:

Create Deny Policy - Files and Folders Scan Results

Create Supplemental Policy - Files and Folders Scan Results

Create Supplemental Policy - Kernel-Mode Policies and Drivers Scan Results

Over time they will be implemented in other list views as well.