AppControl Manager and Harden System Security Update
What's New

Updated the dependencies to the latest versions, including major ones such as .NET SDK and Windows App SDK.

This applies to both Harden System Security and AppControl Manager.
Improved scrolling performance in list views with large data sets by utilizing phased loads based on the importance of each column's data type.

Added icons to many buttons and other areas that didn't have one, helping with discoverability and usability.

Added some icons to the Harden System Security app as well to match the same consistent design.
The Font Families list in the Settings page is now sorted alphabetically, making it easier to find the font name you're looking for.

Improved localization.

Disabled sticky headers in installed apps lists in Create Deny Policy and Create Supplemental Policy pages to better utilize the available vertical space.

The Get Code Integrity Hashes page now supports dragging files onto it and dropping them there to calculate their hashes. (Available only when the app running unelevated)

In the Simulation page, replaced the "Save output to CSV" option with "Export to JSON". JSON is a more structured format and exporting to JSON is a lot faster due to source generated serializers. This also aligns with other export options in the app because they all use JSON as the output format.

Improved the formatting of the exported JSON data from list views. They are now all pretty printed with indentation for better readability.

🎊 Whenever you select an item by browsing for it, the button responsible for it will have a distinct shadow around it and its text will be changed to "Selected" (localized), indicating that you've selected that item. Completes this feature request. As a reminder, you can always review the items you've selected and clear your selections by right-clicking or tap + holding on the button.

Improved the design of UI elements that show the total count of items in each list view. They now have a more sophisticated and consistent design across the entire app and show the count of the data instantly without any delay, so if there was even a few milliseconds delay before, it's not there anymore.

This applies to both Harden System Security and AppControl Manager.
Activated additional thorough analyzers for the Harden System Security app, enabling early detection of more possible issues during compilation.

Aligned the update availability experience in the Harden System Security app with the one in AppControl Manager. Now, when an update is available, you'll see a small icon on the lower left corner on the Update page's icon, informing you that there is a new version available. As a reminder, you can turn off auto-update check in the Update page by simply toggling a button.

When using the Save feature in the Policy Editor, the version number of the policy is contextually incremented. Completes this feature request.